+++ src/main/java/org/jolokia/docker/maven/assembly/DockerFileBuilder.java
@@ -64,7 +64,7 @@ public class DockerFileBuilder {
 
         // Entries
         for (AddEntry entry : addEntries) {
-            b.append("ADD ").append(entry.source).append(" ")
+            b.append("COPY ").append(entry.source).append(" ")
              .append(exportDir).append("/").append(entry.destination).append("\n");
         }
 
@@ -78,7 +78,10 @@ public class DockerFileBuilder {
         }
 
         // Volume export
-        b.append("VOLUME [\"").append(exportDir).append("\"]\n");
+        if (!"/".equalsIgnoreCase(exportDir)) {
+            // Dont export a top level export dir
+            b.append("VOLUME [\"").append(exportDir).append("\"]\n");
+        }
 
         // Environment variable support
         for (Map.Entry<String,String> entry : envEntries.entrySet()) {